Cells Know where there neighbours are now
I wrote a function that found the surrounding cells of the current cell and applyed them to a jagged array of map, the map array was originally a regular array and i need to change it to a jagged for the function. This will allow more functions to be rewriten better.